+++ src/components/pages/apps/matcalc/toolbars/home-toolbar.tsx
@@ -71,8 +71,11 @@ export function HomeToolbar() {
         <ToolbarOpenFile
           onClick={() => {
             openFilesDialog({
-              onFileChange: (message, files) => {
-                onTextFileChange(message, files, (files) => {
+              onFileChange: (files) => {
+                onTextFileChange(files, ({ success, files }) => {
+                  if (!success) {
+                    return
+                  }
                   openMatcalcDialog({
                     type: 'open-table-file',
                     payload: { files, callback: openDataFrames },