+++ src/components/pages/open-files.tsx
@@ -113,7 +113,7 @@ export function arrayBufferToBase64(buffer: ArrayBuffer): string {
 interface IProps {
   message?: string
   //onOpenChange?: (message: string) => void
-  onFileChange?: (message: string, files: FileList | []) => void
+  onFileChange?: (files: FileList | [], message: string) => void
   dirMode?: boolean
   multiple?: boolean
   fileTypes?: string[] | undefined
@@ -144,12 +144,12 @@ export function openFilesDialog({
   input.accept = !dirMode ? getFileTypes(fileTypes) : ''
 
   input.onchange = () => {
-    onFileChange?.(message, input.files ?? [])
+    onFileChange?.(input.files ?? [], message)
     input.remove()
   }
 
   input.oncancel = () => {
-    onFileChange?.(message, [])
+    onFileChange?.([], message)
     input.remove()
   }
 
@@ -268,14 +268,18 @@ export async function readFiles(files: FileList | File[]): Promise<string[]> {
   return fileContents // Return an array with all file contents
 }
 
+type TextFileResult = {
+  success: boolean
+  message?: string
+  files: ITextFileOpen[]
+}
+
 export async function onTextFileChange(
-  _message: string,
   files: File[] | FileList | null,
-  onSuccess: (files: ITextFileOpen[]) => void,
-  onFailure?: (message: string, files: File[] | FileList | null) => void
+  result: (res: TextFileResult) => void
 ) {
   if (!files || files.length === 0) {
-    onFailure?.('no files', files)
+    result({ success: false, message: 'no files', files: [] })
     return
   }
 
@@ -290,7 +294,25 @@ export async function onTextFileChange(
     ext: file.name.split('.').pop() || '',
   }))
 
-  onSuccess(ret)
+  result({ success: true, message: '', files: ret })
+}
+
+export async function onTextFilePaste(
+  text: string,
+  result: (res: TextFileResult) => void
+) {
+  if (!text) {
+    result({ success: false, message: 'no files', files: [] })
+    return
+  }
+
+  const file = {
+    name: 'pasted.txt',
+    text,
+    ext: 'txt',
+  }
+
+  result({ success: true, message: '', files: [file] })
 }
 
 export function onBinaryFileChange(